+++ Stepic/Sources/Modules/Settings/SettingsProvider.swift
@@ -22,8 +22,6 @@ protocol SettingsProviderProtocol: AnyObject {
 
     var isAutoplayEnabled: Bool { get set }
     var isAdaptiveModeEnabled: Bool { get set }
-
-    func deleteAllDownloadedContent() -> Promise<Void>
 }
 
 final class SettingsProvider: SettingsProviderProtocol {
@@ -35,9 +33,6 @@ final class SettingsProvider: SettingsProviderProtocol {
     private let autoplayStorageManager: AutoplayStorageManagerProtocol
     private let adaptiveStorageManager: AdaptiveStorageManagerProtocol
     private let applicationThemeService: ApplicationThemeServiceProtocol
-
-    private let downloadsProvider: DownloadsProviderProtocol
-    private let arQuickLookStoredFileManager: ARQuickLookStoredFileManagerProtocol
 
     var globalDownloadVideoQuality: DownloadVideoQuality {
         get {
@@ -129,9 +124,7 @@ final class SettingsProvider: SettingsProviderProtocol {
         stepFontSizeStorageManager: StepFontSizeStorageManagerProtocol,
         autoplayStorageManager: AutoplayStorageManagerProtocol,
         adaptiveStorageManager: AdaptiveStorageManagerProtocol,
-        applicationThemeService: ApplicationThemeServiceProtocol,
-        downloadsProvider: DownloadsProviderProtocol,
-        arQuickLookStoredFileManager: ARQuickLookStoredFileManagerProtocol
+        applicationThemeService: ApplicationThemeServiceProtocol
     ) {
         self.downloadVideoQualityStorageManager = downloadVideoQualityStorageManager
         self.streamVideoQualityStorageManager = streamVideoQualityStorageManager
@@ -141,18 +134,5 @@ final class SettingsProvider: SettingsProviderProtocol {
         self.autoplayStorageManager = autoplayStorageManager
         self.adaptiveStorageManager = adaptiveStorageManager
         self.applicationThemeService = applicationThemeService
-        self.downloadsProvider = downloadsProvider
-        self.arQuickLookStoredFileManager = arQuickLookStoredFileManager
-    }
-
-    func deleteAllDownloadedContent() -> Promise<Void> {
-        firstly {
-            self.downloadsProvider.fetchCachedCourses()
-        }.then { cachedCourses in
-            self.downloadsProvider.deleteCachedCourses(cachedCourses)
-        }.then { _ -> Promise<Void> in
-            try? self.arQuickLookStoredFileManager.removeAllARQuickLookStoredFiles()
-            return Promise.value(())
-        }
     }
 }
