+++ Stepic/Sources/Services/Common/ApplicationThemeService.swift
