+++ Stepic/Sources/Services/Common/Spotlight/SpotlightContinueUserActivityService.swift
