+++ Stepic/Sources/Services/Common/Spotlight/SpotlightIndexingService.swift
